requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-bower==5.2.0
django-filter==22.1
django-multiselectfield==0.1.12
django-scheduler==0.10.0
fontawesomefree==6.2.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
icalendar==5.0.0
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==9.2.0
protobuf==4.21.7
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
